feat: implement DB metadata storage via sqlmodel
To soon store data not only in memory but also persist it on a more
persistent medium, we replace the global store variable with a SQL based
metadata storage to store and retrieve data.

Right now, this drops all data after a restart, proper persistency will
come after migration to postgres and setting up migrations of db
schemes.

Signed-off-by: Stefan Kraus <dev@stefankraus.org>