feat: return OK on existing upload
As 409 conflict is more expressing that there is a conflict state
happening while processing, and not an already existing upload from
the past one forgot about, we decide to have the same behavior like
the upload_id to work with that. This is a more user friendly behavior
than enforcing to search for it manually.

Signed-off-by: Stefan Kraus <dev@stefankraus.org>